app/middleware/pyproject.toml
[project]
name = "student-lab-mw"
version = "0.1.0"
description = "Middleware for Student Lab Project"
authors = [
    {name = "the1bit",email = "tibor.kiss@cloudmentor.hu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "structlog (>=25.1.0,<26.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "prometheus-fastapi-instrumentator (>=7.0.2,<8.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
